fix text_materials import
git-svn-id: https://dev.cedis.fu-berlin.de/svn/eaz/zwar_archive/trunk@3549 34bfa2f4-0c52-0410-afca-732f319d8f5a

+++ app/models/text_material.rb
@@ -31,7 +31,7 @@ class TextMaterial < ActiveRecord::Base
   def document_file_name=(filename)
     # assign the document - but skip this part on subsequent changes of the file name
     # (because the filename gets assigned in the process of assigning the file)
-    return if filename.blank? || !File.exists?(filename)
+    return if filename.blank?
     filename = (filename || '').sub!(/\w{3,4}$/,'pdf')
     if !defined?(@assigned_filename) || @assigned_filename != filename
       archive_id = ((filename || '')[/^za\d{3}/i] || '').downcase
@@ -56,6 +56,7 @@ class TextMaterial < ActiveRecord::Base
         end
         doc_path
       end
+      return unless File.exists?(filepath)
       File.open(filepath, 'r') do |file|
         self.document = file
       end
